[IntroCSPy-week5] - working on cleaing up some functions and files
